+++ src/main/java/com/example/walletapplication/repository/IntraTransactionRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface IntraTransactionRepository extends JpaRepository<IntraTransaction, Long> {
-    List<IntraTransaction> findByWallets(List<Long> wallets);
+public interface IntraTransactionRepository extends JpaRepository<IntraTransaction, Integer> {
+    List<IntraTransaction> findByWallets(List<Integer> wallets);
 
-    List<IntraTransaction> findByWalletsAndDate(List<Long> wallets, LocalDateTime startDate, LocalDateTime endDate);
+    List<IntraTransaction> findByWalletsAndDate(List<Integer> wallets, LocalDateTime startDate, LocalDateTime endDate);
 }
